+++ src/main/java/umc/nook/users/oauth/KakaoResponseParams.java
@@ -1,0 +1,2 @@
+package umc.nook.users.oauth;public class KakaoResponseParams {
+}
